14.09.2021

+++ src/main/resources/static/js/notesPage.js
@@ -4,14 +4,22 @@ function addNewElement(note) {
 	let noteElement = document.createElement("div")
 	let titleElement = document.createElement("a")
 	titleElement.innerText = note.title
-	titleElement.href = note.links[0].href
+	if (note.hasOwnProperty("links")) {
+		titleElement.href = note.links[0].href
+	}
+	if (note.hasOwnProperty("_links")) {
+		titleElement.href = note._links.get.href
+	}
 	let descriptionElement = document.createElement("div")
 	descriptionElement.innerText = note.description
 	let deleteButton = document.createElement("button")
 	deleteButton.type = "button"
 	deleteButton.innerText = "Delete"
 	deleteButton.onclick = async function() {
-		await deleteNote(note)
+		if (confirm("Delete this note?")) {
+			await deleteNote(note)
+			rootElement.removeChild(noteElement)
+		}
 	}
 	noteElement.appendChild(titleElement)
 	noteElement.appendChild(descriptionElement)
